+++ arbres/src/arbres.c
@@ -18,7 +18,7 @@ int est_vide(arbre a){
 }
 
 // Crée un noeud avec ses caractéristiques
-arbre creer_noeud(int numlex, int numdecl, int nature, int val_e, int val_r){
+arbre creer_noeud(int numlex, int numdecl, int nature, int val_e, double val_r){
   arbre a = malloc(sizeof(struct arbre));
 
   a->numlex = numlex;
